Remove binary files: src part
Remove code which is not usable without hdcp.bin file

https://github.com/ARM-software/tf-issues/issues/651
It is only needed for HDCP support (a DRM feature for video streaming),
if you don't care about that the rest of the firmware should work
perfectly fine without it. Just change rk3399/plat_sip_calls.c to not
handle the RK_SIP_HDCP* SMCs when this option is set, and then exclude
rk3399/drivers/dp/cdn_dp.c and the blob from rk3399/platform.mk.